+++ src/test/java/stream/alwaysbecrafting/flare/GameEngineTest.java
@@ -56,7 +56,7 @@ public class GameEngineTest {
 		GameEngine engine = new GameEngine();
 		StringBuilder output = new StringBuilder( "" );
 		GameSystem system = new GameSystem() {
-			@Override public void onUpdate( GameEngine engine, float deltaTime ) {
+			@Override public void onUpdate( GameEngine engine, double deltaTime ) {
 				output.append( "executed" );
 			}
 		};
@@ -76,7 +76,7 @@ public class GameEngineTest {
 		GameEngine engine = new GameEngine();
 		StringBuilder output = new StringBuilder( "" );
 		GameSystem system = new GameSystem() {
-			@Override public void onUpdate( GameEngine engine, float deltaTime ) {
+			@Override public void onUpdate( GameEngine engine, double deltaTime ) {
 				output.append( "executed" );
 			}
 		};
@@ -97,27 +97,27 @@ public class GameEngineTest {
 		StringBuilder output = new StringBuilder( "" );
 
 		GameSystem system1 = new GameSystem() {
-			@Override public void onUpdate( GameEngine engine, float deltaTime ) {
+			@Override public void onUpdate( GameEngine engine, double deltaTime ) {
 				output.append( "1" );
 			}
 		};
 		GameSystem system2 = new GameSystem() {
-			@Override public void onUpdate( GameEngine engine, float deltaTime ) {
+			@Override public void onUpdate( GameEngine engine, double deltaTime ) {
 				output.append( "2" );
 			}
 		};
 		GameSystem system3 = new GameSystem() {
-			@Override public void onUpdate( GameEngine engine, float deltaTime ) {
+			@Override public void onUpdate( GameEngine engine, double deltaTime ) {
 				output.append( "3" );
 			}
 		};
 		GameSystem system4 = new GameSystem() {
-			@Override public void onUpdate( GameEngine engine, float deltaTime ) {
+			@Override public void onUpdate( GameEngine engine, double deltaTime ) {
 				output.append( "4" );
 			}
 		};
 		GameSystem system5 = new GameSystem() {
-			@Override public void onUpdate( GameEngine engine, float deltaTime ) {
+			@Override public void onUpdate( GameEngine engine, double deltaTime ) {
 				output.append( "5" );
 			}
 		};
@@ -181,17 +181,17 @@ public class GameEngineTest {
 		StringBuilder builder = new StringBuilder( "" );
 
 		GameSystem system1 = new GameSystem() {
-			@Override public void onUpdate( GameEngine engine, float deltaTime ) {
+			@Override public void onUpdate( GameEngine engine, double deltaTime ) {
 				builder.append( "[1]" );
 			}
 		};
 		GameSystem system2 = new GameSystem() {
-			@Override public void onUpdate( GameEngine engine, float deltaTime ) {
+			@Override public void onUpdate( GameEngine engine, double deltaTime ) {
 				builder.append( "[2]" );
 			}
 		};
 		GameSystem system3 = new GameSystem() {
-			@Override public void onUpdate( GameEngine engine, float deltaTime ) {
+			@Override public void onUpdate( GameEngine engine, double deltaTime ) {
 				builder.append( "[3]" );
 			}
 		};
